Update and rename nitclk_master_trigger.py to nitclk_synchronize_trigger.py
