--- src/nitclk/examples/nitclk_master_trigger.py
@@ -1,51 +1,0 @@
-import argparse
-import niscope
-import nitclk
-import sys
-
-
-def example(resource_name1, resource_name2, channels, options, length, voltage):
-    with niscope.Session(resource_name=resource_name1, options=options) as session1, niscope.Session(resource_name=resource_name2, options=options) as session2:
-        session_list = [session1, session2]
-        for session in session_list:
-            session.configure_vertical(range=voltage, coupling=niscope.VerticalCoupling.DC)
-            session.configure_horizontal_timing(min_sample_rate=50000000, min_num_pts=length, ref_position=50.0, num_records=1, enforce_realtime=True)
-        session1.trigger_type = niscope.TriggerType.SOFTWARE
-        nitclk.configure_for_homogeneous_triggers(session_list)
-        nitclk.synchronize(session_list, 200e-9)
-        nitclk.initiate(session_list)
-        session1.send_software_trigger_edge(niscope.WhichTrigger.START)
-        waveforms = session2.channels[channels].fetch(num_samples=length)
-        for i in range(len(waveforms)):
-            print('Waveform {0} information:'.format(i))
-            print(str(waveforms[i]) + '\n\n')
-
-
-def _main(argsv):
-    parser = argparse.ArgumentParser(description='Acquires one record from the given channels.', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('-n1', '--resource-name1', default='PXI1Slot2', help='Resource name of a National Instruments Digitizer')
-    parser.add_argument('-n2', '--resource-name2', default='PXI1Slot3', help='Resource name of a National Instruments Digitizer')
-    parser.add_argument('-c', '--channels', default='0', help='Channel(s) to use')
-    parser.add_argument('-l', '--length', default=1000, type=int, help='Measure record length')
-    parser.add_argument('-v', '--voltage', default=1.0, type=float, help='Voltage range (V)')
-    parser.add_argument('-op', '--option-string', default='', type=str, help='Option string')
-    args = parser.parse_args(argsv)
-    example(args.resource_name1, args.resource_name2, args.channels, args.option_string, args.length, args.voltage)
-
-
-def main():
-    _main(sys.argv[1:])
-
-
-def test_example():
-    options = {'simulate': True, 'driver_setup': {'Model': '5164', 'BoardType': 'PXIe', }, }
-    example('PXI1Slot2', 'PXI1Slot13', '0', options, 1000, 1.0)
-
-
-def test_main():
-    cmd_line = ['--option-string', 'Simulate=1, DriverSetup=Model:5164; BoardType:PXIe', ]
-    _main(cmd_line)
-
-
-if __name__ == '__main__':
-    main()
